Rearchitecting in progress -- Just loads a blank screen right now, got bare-bones class structure and loading system working

+++ src/entities/city.rs
@@ -1,10 +1,9 @@
-#[macro_use]
 use std::fs::File;
 use std::io::{BufRead, BufReader};
 
 use cgmath::Rad;
 
-use coords::LatLong;
+use super::coords::LatLong;
 
 pub struct City {
 	pub name: String,
@@ -14,11 +13,13 @@ pub struct City {
 }
 impl std::fmt::Debug for City {
 	fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-		write!(
-			f,
-			"{}: pop {}, country {}, lat {:?}, long {:?}",
-			self.name, self.population, self.country, self.coords.lat, self.coords.long
-		)?;
+		write!(f,
+		       "{}: pop {}, country {}, lat {:?}, long {:?}",
+		       self.name,
+		       self.population,
+		       self.country,
+		       self.coords.lat,
+		       self.coords.long)?;
 		Ok(())
 	}
 }
@@ -36,10 +37,8 @@ pub fn read_cities_file(file_name: &str) -> Vec<City> {
 			name: String::from(columns[0]),
 			population: columns[1].parse::<u32>().unwrap(),
 			country: String::from(columns[2]),
-			coords: LatLong::new(
-				Rad(columns[3].parse::<f32>().unwrap()),
-				Rad(columns[4].parse::<f32>().unwrap()),
-			),
+			coords: LatLong::new(Rad(columns[3].parse::<f64>().unwrap()),
+			                     Rad(columns[4].parse::<f64>().unwrap())),
 		};
 		cities.push(city);
 	}
